Updated TweenCamera to use double values && added option to pass method to receive controller on webview created

+++ threeJS_Viewer/lib/threeJSController.dart
@@ -57,7 +57,7 @@ class ThreeJSController {
     webController?.runJavascript('window.resetCameraControls(${autoRotate})');
   }
 
-  void tweenCamera(int targetX, int targetY, int targetZ, int duration, bool autoRotate) {
+  void tweenCamera(double targetX, double targetY, double targetZ, double duration, bool autoRotate) {
     webController?.runJavascript('window.tweenCamera(${targetX}, ${targetY}, ${targetZ}, ${duration})');
     if (Platform.isIOS) resetCameraControls(autoRotate);
   }
